Add Default impl for JitterRange
